gandra-tools-app/gandra-tools-api/pyproject.toml
[project]
name = "gandra-tools"
version = "0.1.0"
description = "Swiss-army toolset — YouTube transcripts, RAG research, file ops, image OCR, and more"
requires-python = ">=3.11"
dependencies = [
    # FastAPI
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    # CLI
    "typer>=0.15",
    "rich>=13.0",
    # Auth
    "python-jose[cryptography]>=3.3",
    # LLM providers
    "openai>=1.60",
    "anthropic>=0.42",
    "httpx>=0.28",
    # Database
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "greenlet>=3.0",
    # Publisher
    "jinja2>=3.1",
    # Utilities
    "python-slugify>=8.0",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
youtube = [
    "youtube-transcript-api>=0.6",
    "yt-dlp>=2024.0",
]
imageops = [
    "easyocr>=1.7",
    "Pillow>=10.0",
    "opencv-python>=4.8",
]
research = [
    "beautifulsoup4>=4.12",
    "trafilatura>=1.6",
]
all = [
    "gandra-tools[youtube,imageops,research]",
]

[project.scripts]
gandra-tools = "gandra_tools.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gandra_tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
]
